% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a104apps.R
\name{a104apps}
\alias{a104apps}
\title{Pick an interative shiny app to launch}
\usage{
a104apps()
}
\description{
Running this function produces a simple menu. Enter the number corresponding to the interactive app you want to launch.
}
\examples{
\dontrun{
a104apps()
}
}
