% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirtToClassical.R
\name{alfunc}
\alias{alfunc}
\title{Cronbach's alpha.}
\usage{
alfunc(data)
}
\arguments{
\item{data}{A data frame of item data.}
}
\description{
This function estimates Cronbach's alpha from a data set of item score. For the purposes of this function only
any missing items are rescored to zero.
}
