% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentdata.R
\name{augmentdata}
\alias{augmentdata}
\title{Augment existing item data}
\usage{
augmentdata(data)
}
\arguments{
\item{data}{A data frame containing information about all the items.}
}
\value{
A data frame with additional rows of augmented data (if necessary) and an attribute "fakedata" indicating the rows in the data frame that have been added in.
}
\description{
This function first checks if there are any missing categories between 0 and the max available for each item in a data set.
For each missing category, an additional row is added to the data with missing values for all items
except the one with the missing category and with the value for this item set equal to the missing category.
Using augmented data provides an alternative to the default behaviour of "mirt" where items are entirely rescored
to remove missing categories. The augmented data approach can help facilitate later analyses such as selecting
a number of items with an overall total number of marks fixed at some level (see runItemSelectionApp). Provided
that a reasonably large amount of data is being analysed and very little data is being added to augment the data set,
this step should have minimal impact upon estimated item parameters.
}
\examples{
aug1=augmentdata(mathsdata[1:10,15:20])
aug1
attributes(aug1)
}
