% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSEM_scale_function.R
\name{CSEM_scale_function}
\alias{CSEM_scale_function}
\title{Calculate the Conditional Standard Error of Measurement (CSEM) of scale scores}
\usage{
CSEM_scale_function(jointdist, my_scale_func)
}
\arguments{
\item{jointdist}{A data.frame giving the joint distribution of true and observed raw scores into the format created by 'JointAbilityScoreDist'.}

\item{my_scale_func}{A function that can taken any valid vector of raw scores and return the transformed scores on the reporting scale.}
}
\value{
The function returns a data.frame with the following columns
\describe{
  \item{theta}{Value of abilit}
  \item{p_theta}{Marginal distribution of theta}
  \item{true_raw}{True score associated with value of ability on raw score scale}
  \item{true_scale}{True score associated with value of ability on transformed reporting scale}
  \item{csem_raw}{CSEM on scale of raw scores}
  \item{csem_scale}{CSEM on scale of transformed scores on the reporting scale}
  }
}
\description{
The function is used to calculate the CSEM of some conversion of raw scores to a different reporting scale (e.g., coarse grades).
}
\examples{
\dontrun{
#library(unimirt)
mirtRasch=unimirt(mathsdata,"Rasch")
jdist=JointAbilityScoreDist(mirtRasch)
csem1=CSEM_scale_function(jdist,I)#if transformation is just the identity function
head(csem1)
ggplot(data=csem1,aes(x=true_raw,y=csem_raw))+geom_line()
ggplot(data=csem1,aes(x=true_scale,y=csem_scale))+geom_line()

#a more interesting example
#imagine raw scores are converted to grades from 0 to 5 according to some cut-scores
scale_func=stepfun(x=c(17,33,47,60,72),y=0:5)
csem2=CSEM_scale_function(jdist,scale_func)
ggplot(data=csem2,aes(x=true_raw,y=csem_raw))+geom_line()
ggplot(data=csem2,aes(x=true_raw,y=csem_scale))+geom_line()
ggplot(data=csem2,aes(x=true_scale,y=csem_scale))+geom_line()
}
}
