% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSEMfromMirt.R
\name{CSEMfromMirt}
\alias{CSEMfromMirt}
\title{Function to calculate conditional standard error of measurement at each true score}
\usage{
CSEMfromMirt(mirtobj, which.items = NULL)
}
\arguments{
\item{mirtobj}{An estimated IRT model (of class SingleGroupClass) estimated either using \link[mirt]{mirt} or \link[unimirt]{unimirt}. 
Must be fitted using the (default) "EM" method. Must be unidimensional.}

\item{which.items}{A vector of denoting which items should be included in calculating the score.}
}
\value{
The function returns a data.frame with columns: true.score (a vector of possible integer true scores) 
, and csem (the standard deviation of obsereved scores for candidates with the given true score).
}
\description{
Calculates the conditional standard error of measurement at each unit true score
on the test (totalled across a given selection of items).
This is defined as the standard deviation of observed scores given the true score.
}
\examples{
\dontrun{
mirt1=unimirt(mathsdata,"2")
CSEMfromMirt(mirt1)
}
}
