% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DichotomousAndPolytomous.R
\name{dichotomous.to.pcm}
\alias{dichotomous.to.pcm}
\title{Convert Rasch difficulties of many dichotomous items to Partial Credit Model (PCM) thresholds}
\usage{
dichotomous.to.pcm(dich_diffs)
}
\arguments{
\item{dich_diffs}{A vector of Rasch difficulties. The length of the vector represents the number of dichotomous items considered.}
}
\value{
A vector of thresholds for the partial credit model (PCM).
}
\description{
Given a set of Rasch difficulties from n dichomtomous items, calculate the equivalent thresholds for a single polytomous item
(with n+1 categories) that fits the partial credit model (PCM).
}
\examples{
\dontrun{
dichotomous.to.pcm(c(0,0))
pcm.to.dichotomous(c(-0.6931472,0.6931472))

dichotomous.to.pcm(c(-2,-4,1))
pcm.to.dichotomous(c(-4.132845,-1.922140,1.054985))

dichotomous.to.pcm(c(-1,-1,1,1))
pcm.to.dichotomous(c(-1.8200752,-0.6243906,0.6243906,1.8200752))
}
}
