% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmpiricalICCfit.R
\name{EmpiricalICCfit}
\alias{EmpiricalICCfit}
\title{Create a ggplot of empirical item characteristic curves compared to theoretical relationship based on the model}
\usage{
EmpiricalICCfit(mirtobj, itenum, which.items = NULL, ngroups = 1)
}
\arguments{
\item{mirtobj}{An estimated IRT model (of class SingleGroupClass) estimated either using \link[mirt]{mirt} or \link[unimirt]{unimirt}.}

\item{itenum}{A numeric input denoting which item the plot should be produced for.}

\item{which.items}{A numeric vector denoting which items should be used to create the total test score. By default all available items are included.}

\item{ngroups}{The number of groups to split the cohort into in order to produce the empirical points on the curve (the mean total score in each group is plotted against the mean item score). Setting ngroups=1 (the default and the exception to the usual pattern) will create one group for each possible total test score.}
}
\value{
A list with the following elements.
\describe{
  \item{plot1}{A function that translates any vector of scores on form X into equivalent scores on form Y.}
  \item{modelchartdat}{A data frame giving the expected relationship between total score and mean item score based on the IRT model}
  \item{empiricalchartdat}{A data frame giving the empirical results within each group}
}#'
}
\description{
The empirical item characteristic curve plots the mean scores on each item for groups with different total scores
on the whole test.
}
\examples{
\dontrun{
mirt1=mirt(mathsdata,1)
EICC=EmpiricalICCfit(mirt1,1)
EICC
EICC$plot1
EmpiricalICCfit(mirt1,1,ngroups=10)$plot1
}
}
