% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimatedItemDistribution.R
\name{EstimatedItemDistribution}
\alias{EstimatedItemDistribution}
\title{Estimate the score distribution for an individual item in a "mirt" object.}
\usage{
EstimatedItemDistribution(mirtobj, which.item, theta = NULL, qwts = NULL)
}
\arguments{
\item{mirtobj}{An estimated IRT model (of class SingleGroupClass) estimated either using \link[mirt]{mirt} or \link[unimirt]{unimirt}.}

\item{which.item}{An integer denoting which item to investigate.}

\item{theta}{A matrix giving the ability values used in estimation. By default this is extracted directly from mirtobj.}

\item{qwts}{A vector giving the weight to assign to each ability in thetas. By default this is extracted directly from mirtobj.}
}
\description{
This function uses the fitted model estimates the percentage of persons that will get each score.
If there is no missing data for the item (i.e. all persons have atctually taken the item)
then the function will also provide the actual percentages.
}
\examples{
\dontrun{
mirt1=unimirt(mathsdata,"2")
EstimatedItemDistribution(mirt1,1)
EstimatedItemDistribution(mirt1,19)
}
}
