% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirtStockingLordRecalibrate.R
\name{ExpectedScoreCompare}
\alias{ExpectedScoreCompare}
\title{Plot comparing expected total scores on common items between mirt objects}
\usage{
ExpectedScoreCompare(mirtobj1, mirtobj2, lab1 = NULL, lab2 = NULL)
}
\arguments{
\item{mirtobj1}{An estimated IRT model (of class SingleGroupClass) estimated either using the function "unimirt"
or by applying the function "mirt" directly.}

\item{mirtobj2}{An estimated IRT model (of class SingleGroupClass) estimated either using the function "unimirt"
or by applying the function "mirt" directly.}

\item{lab1}{Optional character string that can be added to change labelling of object 1 in the chart.}

\item{lab2}{Optional character string that can be added to change labelling of object 2 in the chart.}
}
\description{
Common items are identified automatically on the basis of shared variable names.
}
\examples{
\dontrun{
mirt1=unimirt(mathsdata[1:250,1:10])
mirt2=unimirt(mathsdata[251:500,6:15])
ExpectedScoreCompare(mirt1,mirt2)

mirt2.rescale=MirtObjectRecalibrate(mirt1,mirt2)
ExpectedScoreCompare(mirt1,mirt2.rescale)
}
}
