% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindBestRabils.R
\name{FindBestRabils}
\alias{FindBestRabils}
\title{Function to estimate how big the R_rest values for a set of items might get}
\usage{
FindBestRabils(mirtobj)
}
\arguments{
\item{mirtobj}{An estimated IRT model (of class SingleGroupClass) estimated either using the function "unimirt"
or by applying the function "mirt" directly.}
}
\value{
A vector of highest possible correlations between true ability and item scores.
}
\description{
For each item in an IRT model in turn, this function estimates the largest possible theoretical correlation between
true ability and item scores that might occur if the only thing that changed was the mean ability of the population
(i.e. item parameters as well as standard deviation and shape of ability distribution are fixed). This allows an
idea of how high an R_rest value might get if the item was used in a population suitable to its difficulty
and if the scores on the remainder of the test were perfectly reliable.
}
\details{
For the purposes of this function only the underlying ability distribution is assumed to be normally distributed.
Also note that R_abil is actually estimated as the square root of the percentage of the variance in item scores
that can be explained by ability. As such, it accounts for nonlinear relationships between item scores and ability.
}
\examples{
\dontrun{
mirt1=unimirt(mathsdata,"2")
FindBestRabils(mirt1)
}
}
