% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDataFromMirt.R
\name{GetDataFromMirt}
\alias{GetDataFromMirt}
\title{Function to extract the (non-fake) item data set from a mirt object}
\usage{
GetDataFromMirt(mirtobj)
}
\arguments{
\item{mirtobj}{An estimated IRT model (of class SingleGroupClass) estimated either using the function "unimirt"
or by applying the function "mirt" directly.}
}
\value{
A matrix of item data.
}
\description{
Extracts the data frame used for model fitting from within a 'mirt' object.
This function will automatically remove any rows of fake data added by the augmentdata function.
}
\examples{
\dontrun{
smallmirt=unimirt(mathsdata[1:10,15:20],"Rasch")
smallmirt@Data$data
GetDataFromMirt(smallmirt)
}
}
