% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlternativeThresholds.R
\name{InformationLocation}
\alias{InformationLocation}
\title{Location of each item's maximum information}
\usage{
InformationLocation(mirtobj, which.items = NULL)
}
\arguments{
\item{mirtobj}{An estimated IRT model (of class SingleGroupClass) estimated using \link[mirt]{mirt} or \link[unimirt]{unimirt}.}

\item{which.items}{an integer vector indicating which items to include. By default all items are included.}
}
\value{
A data frame of item names, item maxima and theta locations of maximum information.
}
\description{
This function calculates the ability value at which each item has the highest value for its information function.
For dichotomous items this will usually be the same as the item difficulty parameter.
For polytomous items it may provide an idea of the point in the ability scale where the item is most useful.
}
\examples{
\dontrun{
#model using "Science" data from mirt package
scimod=unimirt(Science[,1:4]-1,"gpcm")
InformationLocation(scimod)
#visual verification
unimirt.plot(scimod,"infotrace")
}
}
