% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirtStockingLordRecalibrate.R
\name{ItemParameterCompare}
\alias{ItemParameterCompare}
\title{Plot comparing item parameters for common items}
\usage{
ItemParameterCompare(mirtobj1, mirtobj2, compare = "Difficulties")
}
\arguments{
\item{mirtobj1}{An estimated IRT model (of class SingleGroupClass) estimated either using the function "unimirt"
or by applying the function "mirt" directly.}

\item{mirtobj2}{An estimated IRT model (of class SingleGroupClass) estimated either using the function "unimirt"
or by applying the function "mirt" directly.}

\item{compare}{Character string equal to "Difficulties" (the default) or "Slopes" defining which item parameters should be compared.}
}
\description{
Common items are identified automatically on the basis of shared variable names.
}
\examples{
\dontrun{
mirt1=unimirt(mathsdata[1:250,1:10])
mirt2=unimirt(mathsdata[251:500,6:15])

ItemParameterCompare(mirt1,mirt2, compare = "Difficulties")
ItemParameterCompare(mirt1,mirt2, compare = "Slopes")
}
}
