% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointAbilityScoreDist.R
\name{JointAbilityScoreDist}
\alias{JointAbilityScoreDist}
\title{Create joint distribution of true scores (i.e., theta) and observed scores}
\usage{
JointAbilityScoreDist(mirtobj, which.items = NULL, theta.mean.sd = NULL)
}
\arguments{
\item{mirtobj}{An estimated IRT model (of class SingleGroupClass) estimated either using \link[mirt]{mirt} or \link[unimirt]{unimirt}. 
Must be fitted using the (default) "EM" method. Must be unidimensional.}

\item{which.items}{A vector of denoting which items should be included in calculating the total raw score. (All items included by default).}

\item{theta.mean.sd}{An optional 2-value vector giving a mean and standard deviation (SD) of ability.
If supplied, this will be used to create 200 quadrature points relevant to the defined mean and SD.
These will be used instead of the quadrature points in the original IRT model.}
}
\value{
The function returns a data.frame with the following columns
\describe{
  \item{theta}{Value of abilit}
  \item{true_score}{True score associated with value of ability}
  \item{raw_score}{Raw (observed) score}
  \item{p}{Joint probability of true score (or theta) and raw observed score occuring together}
  \item{p_theta}{Marginal distribution of theta (will be duplicated for every occurrence of same theta in data)}
  \item{p_raw}{Marginal distribution of raw scores (will be duplicated for every occurrence of same raw score in data)}
  \item{p_raw_given_theta}{Conditional distribution of raw score given theta}
  \item{p_theta_given_raw}{Conditional distribution of theta given raw score}
  \item{e_true}{Expected value of true score given raw score}
  }
}
\description{
Calculate the probability of each combination of true and observed scores.
Note that, since we are dealing with unidimensional models, there is a one-to-one
correspondence between ability values and true scores.
}
\details{
As well as the joint distribution, the output also provides the marginal distribution
of true scores and raw scores derived from on the IRT model.
Marginal distributions of true scores are purely based on the weights provided for each 
quadrature point in model fitting.

Conditional distributions of true scores (i.e., the quadrature points) based on raw scores
and raw score based on true scores are also provided
}
\examples{
\dontrun{
#library(unimirt)
mirtRasch=unimirt(mathsdata,"Rasch")
jdist=JointAbilityScoreDist(mirtRasch)
head(jdist)
plot(jdist$raw_score,jdist$p_raw)
plot(jdist$true_score,jdist$p_theta)
plot(jdist$theta,jdist$true_score)
plot(jdist$raw_score,jdist$e_true)+xlim(0,NA)+ylim(0,NA)+geom_abline()
}
}
