% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirtLogLikMatrix.R
\name{MirtEAPs}
\alias{MirtEAPs}
\title{Function to extract EAP ability estimates from a mirt object}
\usage{
MirtEAPs(mirtobj, excludefake = TRUE)
}
\arguments{
\item{mirtobj}{An estimated unidimensional IRT model (of class SingleGroupClass) estimated either using the function "unimirt"
or by applying the function "mirt" directly.}

\item{excludefake}{A logical value dnoting whether any added "fake" cases should be excluded
from the returned ability estimates (default=TRUE).}
}
\value{
A data frame with column labels "eap" and "se.eap".
}
\description{
A wrapper for the "fscores" function from the package "mirt".
However, by default, this function will ensure any fake data added by the 
link{augmentdata} function as a pre-cursor to model fitting is excluded.
Furthermore, this function automatically includes standard errors for each
estimate and returns the ability estimates and standard errors in a data frame.
}
\examples{
\dontrun{
mirt1=unimirt(mathsdata,"2")
MirtEAPs(mirt1)
}
}
\seealso{
\code{\link[mirt]{fscores}}
}
