% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirtSubset.R
\name{MirtSubset}
\alias{MirtSubset}
\title{Create a mirt object based on a subset of items in another one}
\usage{
MirtSubset(mirtobj, which.items)
}
\arguments{
\item{mirtobj}{An estimated IRT model (of class SingleGroupClass) estimated either using \link[mirt]{mirt} or \link[unimirt]{unimirt}.}

\item{which.items}{an integer vector indicating which items to include in the reduced object.}
}
\value{
A estimated "mirt" object.
}
\description{
This function takes a fitted unidimensional IRT model (fitted using mirt or unimirt) 
as an input, and creates a smaller one based on a particular subset of items.
This may be useful to enable the shiny apps to focus upon a particular subsection of items.
Note that the item parameters are NOT re-estimated. This function is simply intended
to allow users to zoom in on results for particular sets of items.
For example this might be useful to help produce a Wright Map based
on a subset of the items.
}
\details{
Note that although the item parameters are unchanged in this process.
Fit statistics (or anything else dependent upon producing 
ability estimates as an intermediate step) 
will be affected
by focussing on a subset of items.
}
\examples{
\dontrun{
mirt1=unimirt(mathsdata,"Rasch")
MirtTidyCoef(mirt1)
mirt2=MirtSubset(mirt1,c(1,3,5,10))
MirtTidyCoef(mirt2)
}
}
