% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirtTidyCoefSE.R
\name{MirtTidyCoefSE}
\alias{MirtTidyCoefSE}
\title{Get table of item parameters with standard errors}
\usage{
MirtTidyCoefSE(mirtobj)
}
\arguments{
\item{mirtobj}{An estimated IRT model (of class SingleGroupClass) estimated either using \link[mirt]{mirt} or \link[unimirt]{unimirt}.}
}
\value{
A data frame of item parameters and associated standard errors.
}
\description{
This function produces a slightly more condensed table of IRT item parameters than using the "coef" function.
directly on a mirt object. IRT parameters are prefixed by 'a', 'g' or 'b'. Slope parameters are prefixed with an 'a',
difficulty parameters with a 'b' and guessing parameters are labelled 'g'. Standard errors are included in the same
table. Note that if the original IRT model was estimated without the option "SE=TRUE" then this function will give an
error.
}
\examples{
\dontrun{
mirt1=unimirt(mathsdata,"2",SE=TRUE)
MirtTidyCoefSE(mirt1)
}
}
