% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirtToClassical.R
\name{MirtToClassical}
\alias{MirtToClassical}
\title{Get table of classical item statistics for data in a "mirt" object.}
\usage{
MirtToClassical(mirtobj)
}
\arguments{
\item{mirtobj}{An estimated IRT model (of class SingleGroupClass) estimated either using \link[mirt]{mirt} or \link[unimirt]{unimirt}.}
}
\description{
This function extracts the data from a IRT model estimated using "mirt".
Then, if there are no missing item scores, it produces a data frame containing the usual classical test statistics.
}
\examples{
mirt1<-unimirt(mathsdata[1:100,1:10],"Rasch")
MirtToClassical(mirt1)
}
