% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParComparePlot.R
\name{ParComparePlot}
\alias{ParComparePlot}
\title{Plot comparing item parameters}
\usage{
ParComparePlot(
  mirtobjlist,
  mirtnames = NULL,
  commononly = TRUE,
  compare = "Difficulties"
)
}
\arguments{
\item{mirtobjlist}{A list of estimated IRT models. For this function to make
any sense at all these should all be of the same type.}

\item{mirtnames}{A character vector of the same length as "mirtobjlist"
use to add colour labels to the chart. Dy default they are just labelled
obj1, obj2, ...}

\item{commononly}{Logical value denoting whether only items that occur in more
than one object should be included in the chart. Default TRUE.}

\item{compare}{Character string equal to "Difficulties" (the default), "Slopes" or "Guessing" defining
which item parameters should be compared.}
}
\description{
Common items are identified automatically on the basis of shared variable names.
This function differs from "ItemParameterCompare" in that
more than two IRT objects can be compared.
In addition, if all models have been estimated with 'SE=TRUE'
then this function include 95 per cent confidence intervals for
item parameters making it easier to assess whether differences
between item parameters are statistically significant.
}
\examples{
\dontrun{
mirt1=unimirt(mathsdata[1:250,1:10],SE=TRUE)
mirt2=unimirt(mathsdata[251:500,6:15],SE=TRUE)

ParComparePlot(list(mirt1,mirt2),c("first250","next250"), compare = "Difficulties")
ParComparePlot(list(mirt1,mirt2),c("first250","next250"), compare = "Slopes")
}
}
