% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runClassicalApp.R
\name{runClassicalApp}
\alias{runClassicalApp}
\title{Launch interactive shiny app to perform some simple classical item analyses}
\usage{
runClassicalApp(MC = FALSE)
}
\arguments{
\item{MC}{If TRUE launch a version of the App designed for analysis of multiple choice tests.
The only difference is that converting item responses (e.g. "A", "B", "C", "D") to scores
is incorporated in the App and that additional item characteristic curves (ICCs) exploring
the percentage of pupils choosing each category in each score group are included.}
}
\description{
This function launches a shiny app where ".csv" files of data can be uploaded and analysed.
It assumes that all pupils have had a chance to attempt all items.
}
\examples{
\dontrun{
runClassicalApp()

runClassicalApp(MC=TRUE)
}
}
