% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runClassicalEquatingApp.R
\name{runClassicalEquatingApp}
\alias{runClassicalEquatingApp}
\title{Launch interactive shiny app to perform some simple classical equating}
\usage{
runClassicalEquatingApp()
}
\description{
This function launches a shiny app where ".csv" files of data can be uploaded and analysed interactively.
It is nothing to do with IRT and is only including to support a wider training course.
}
\examples{
\dontrun{
runClassicalEquatingApp()
}
}
