% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runItemSelectionApp.R
\name{runItemSelectionApp}
\alias{runItemSelectionApp}
\title{Launch shiny app to facilitate item selection}
\usage{
runItemSelectionApp(interactive.upload = FALSE)
}
\arguments{
\item{interactive.upload}{If TRUE launch a version of the App that allows loading previously fitted IRT models interactively.
Otherwise (default) the Shiny App is purely concerned with viewing results for IRT models
that have already been fitted using R code.}
}
\description{
This function launches a shiny app where allowing item selection based upon information estimated using IRT and stored in "mirt" objects. Supplementary information may also be uploaded to the app to support item selection.
}
\examples{
\dontrun{
runItemSelectionApp()
}
}
