% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMirtCompareApp.R
\name{runMirtCompareApp}
\alias{runMirtCompareApp}
\title{Launch shiny app to compare estimated IRT objects}
\usage{
runMirtCompareApp(interactive.upload = FALSE)
}
\arguments{
\item{interactive.upload}{If TRUE launch a version of the App that allows loading previously fitted IRT models interactively.
Otherwise (default) the Shiny App is purely concerned with viewing results for IRT models
that have already been fitted using R code.}
}
\description{
This function launches a shiny app where allowing a comparison of item parameters and expected total scores on the common items between different estimatyed mirt objects.
}
\examples{
\dontrun{
runMirtCompareApp()
}
}
