% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runResultsApp.R
\name{runResultsApp}
\alias{runResultsApp}
\title{Launch shiny app to view IRT results}
\usage{
runResultsApp()
}
\description{
This function launches a shiny app where IRT objects estimated using "mirt" can be selected and results viewed.
}
\examples{
\dontrun{
runResultsApp()
}
}
