% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreDistFromMirt.R
\name{ScoreDistFromMirt}
\alias{ScoreDistFromMirt}
\title{Function to estimate sum score distribution for a selection of items in a mirt object}
\usage{
ScoreDistFromMirt(mirtobj, which.items = NULL, theta.mean.sd = NULL)
}
\arguments{
\item{mirtobj}{An estimated IRT model (of class SingleGroupClass) estimated either using \link[mirt]{mirt} or \link[unimirt]{unimirt}. 
Must be fitted using the (default) "EM" method. Will work with any form of model (graded response, Rasch,...). Will even work
with multidimensional models but may be innaccurate unless the object incorporates a large number of quadrature points.}

\item{which.items}{A vector of denoting which items should be included in calculating the score.}

\item{theta.mean.sd}{A vector of length 2 giving the mean and standard deviation of the ability distribution. 
If not supplied then these are derived directly from the mirt object. This parameter may only be used for unidimensional models.}
}
\value{
The function returns a data.frame with columns: score (a vector of possible scores) 
,prob (the proportion of candidates expected to achieve each scores), expectedtheta (expected ability given the
total score), and sd theta the standard deviation of abilities for candidates with the given total score.
}
\description{
This function can be used to help facilitate observed score IRT equating.
Early trials suggest this function works best with unidimensional models.
}
\examples{
\dontrun{
mirt1=unimirt(mathsdata,"2")
ScoreDistFromMirt(mirt1)
}
}
