% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tam.jml.to.unimirt.R
\name{tam.jml.to.unimirt}
\alias{tam.jml.to.unimirt}
\title{Convert IRT model fitted using tam.jml (from TAM package) to a mirt object}
\usage{
tam.jml.to.unimirt(tamjml = NULL)
}
\arguments{
\item{tamjml}{.}
}
\description{
This function is designed to allow users to fit a Rasch model using the
joint maximum likelihood (JML) procedure but to still have access to all of the interactive
graphical displays and other functions that are available in unimirt.
}
\examples{
\dontrun{
tamjml=TAM::tam.jml(mathsdata[,1:5])
tamjml$item
mirtfromjml=tam.jml.to.unimirt(tamjml)
MirtTidyCoef(mirtfromjml)
}
}
