% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runResultsApp.R
\name{runResultsApp}
\alias{runResultsApp}
\title{Launch shiny app to view IRT results}
\usage{
runResultsApp(interactive.data = FALSE, WLE = FALSE)
}
\arguments{
\item{interactive.data}{If TRUE launch a version of the App that allows loading data and model fitting also to be interactive.
Otherwise (default) the Shiny App is purely concerned with viewing results for IRT models
that have already been fitted using R code.Only the WLE=FALSE version is available.}

\item{WLE}{If TRUE launch a version of the App (with interactive.data=FALSE)
where all calculations using ability estimates (e.g. fit statistics) are based upon
WLE estimates rather than plausible values (the default). The WLE version tends to run a little more
slowly but may be of interest to some users.}
}
\description{
This function launches a shiny app where IRT objects estimated using "mirt" can be selected and results viewed.
}
\examples{
\dontrun{
runResultsApp()
}
}
